This is the lite preview and mainly used for backup purpose
This will include a full section of the Property section in the loan
applications.

Added a menu button that will be clickable for a user on mobile devices
to jump to any section in the admin

Cleaned up alot of junk code and made adjustments to code to be cleaner.

Restructured the file management

This will be used temp. before make a full preview available.