Completion of section 2 of the loan app
This includes the new section Property and its been completed for all
IDs and Name fields in each input field.

This update also includes the new design for each section to display in
the full loan app the new menu button added.

File structure has been changed and tested successfully as of now.

Links on the left hand side for the menu has been updated too.

Final testing of everything tomorrow to upload to Master