THIS IS THE OFFICIAL COMPANY LIVE DEMO
This will be consider as a full live demo that can be used to demo the
new loan application. This will consist of both access to the full and
the quick loan apps. Users should have the ability to view this loan app
on mobile and desktops with no issues navigating the site.

All changes going forward will be changes made to incorporate the live
version that will be hosted on the server when completed.